connect foros to post
